Update emulator.md
Hi - so I tried my best. Was not able to get through it myself though, so started just doing edits on tone and language. You do need to give it a good read to make sure I did not mess it up, especially the command line step which I took as an option compared to doing it in the Android Studio GUI.